+++ NEWS.md
@@ -58,6 +58,7 @@
 - `$glimpse()` for `DataFrame` has two new arguments `max_items_per_column` and
   `max_colname_length` (#1200).
 - New method `$list$sample()` (#1204).
+- New argument `coalesce` in `$join_asof()` (#1205).
 
 ### Other changes